+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current changes without release yet
 
 - Added support for (min/max)(in/ex)clusive
+- Repaired bug in minLength
 
 ## [v0.1.72] - 2025-06-14